+++ src/teropa-drumkit.d.ts
@@ -1,0 +1,9 @@
+declare module '@teropa/drumkit' {
+  export const hatClosed: string;
+  export const hatClosed2: string;
+  export const hatOpen: string;
+  export const ride: string;
+  export const kick: string;
+  export const snare: string;
+  export const crash: string;
+}